% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFR_analysis.R
\name{plot_lagged_cfr}
\alias{plot_lagged_cfr}
\title{CFR Plotter}
\usage{
plot_lagged_cfr(data, lag, max_lag)
}
\arguments{
\item{data}{The input data}

\item{lag}{The lag}

\item{max_lag}{The maximum lag}
}
\value{
A multi-panel plot of case fatality rates with different lags
}
\description{
Create a multi-panel CFR plotter with different lags
}
